pyproject.toml
[project]
name = "coder-buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq",
    "langchain",
    "langchain-core",
    "langchain-groq",
    "langgraph",
    "pip",
    "pydantic",
    "python-dotenv",
]
